# openapi.yaml
openapi: 3.0.3
info:
  title: Blog API
  version: 1.1.0
  description: API REST para gestionar autores, posts y comentarios.
servers:
  - url: http://localhost:3000
    description: Local
  - url: https://proyectom2-richard-anderson-gonzalez-cantillo-production.up.railway.app
    description: Railway

paths:
  /api/authors:
    get:
      summary: Listar autores
      responses:
        '200':
          description: Lista de autores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
    post:
      summary: Crear autor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorInput'
      responses:
        '201':
          description: Autor creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/authors/{id}:
    get:
      summary: Obtener autor por id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Autor encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar autor
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorUpdate'
      responses:
        '200':
          description: Autor actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Email duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Eliminar autor
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/posts:
    get:
      summary: Listar posts
      parameters:
        - in: query
          name: published
          schema:
            type: boolean
          description: Filtra por publicados (true/false)
      responses:
        '200':
          description: Lista de posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      summary: Crear post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '201':
          description: Post creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/posts/{id}:
    get:
      summary: Obtener post
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Post encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar post
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
      responses:
        '200':
          description: Post actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Eliminar post
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/posts/author/{authorId}:
    get:
      summary: Listar posts por autor
      parameters:
        - in: path
          name: authorId
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Lista de posts del autor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/comments:
    get:
      summary: Listar comentarios
      responses:
        '200':
          description: Lista de comentarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
    post:
      summary: Crear comentario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '201':
          description: Comentario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/comments/{id}:
    get:
      summary: Obtener comentario
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Comentario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar comentario
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentUpdate'
      responses:
        '200':
          description: Comentario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Eliminar comentario
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
        minimum: 1

  responses:
    BadRequest:
      description: Petición inválida (ids no enteros, campos requeridos, tipos incorrectos)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Message:
      type: object
      properties:
        message:
          type: string

    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          description: Código de error opcional (ej. CONFLICT)

    Author:
      type: object
      properties:
        id: { type: integer }
        name: { type: string, maxLength: 100 }
        email: { type: string, format: email }
        bio: { type: string, maxLength: 500 }
        created_at: { type: string, format: date-time }

    AuthorInput:
      type: object
      required: [name, email]
      properties:
        name: { type: string, maxLength: 100 }
        email: { type: string, format: email }
        bio: { type: string, maxLength: 500 }

    AuthorUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 100 }
        email: { type: string, format: email }
        bio: { type: string, maxLength: 500 }

    Post:
      type: object
      properties:
        id: { type: integer }
        title: { type: string, maxLength: 200 }
        content: { type: string, maxLength: 5000 }
        author_id: { type: integer }
        published: { type: boolean }
        created_at: { type: string, format: date-time }

    PostInput:
      type: object
      required: [title, content, author_id]
      properties:
        title: { type: string, maxLength: 200 }
        content: { type: string, maxLength: 5000 }
        author_id: { type: integer }
        published: { type: boolean }

    PostUpdate:
      type: object
      properties:
        title: { type: string, maxLength: 200 }
        content: { type: string, maxLength: 5000 }
        author_id: { type: integer }
        published: { type: boolean }

    Comment:
      type: object
      properties:
        id: { type: integer }
        post_id: { type: integer }
        author_id: { type: integer }
        body: { type: string, maxLength: 1000 }
        created_at: { type: string, format: date-time }

    CommentInput:
      type: object
      required: [post_id, author_id, body]
      properties:
        post_id: { type: integer }
        author_id: { type: integer }
        body: { type: string, maxLength: 1000 }

    CommentUpdate:
      type: object
      properties:
        body: { type: string, maxLength: 1000 }